requirements.txt
python-dotenv==1.0.0
pydantic==2.5.2
pandas==2.0.3
numpy==1.26.4
python-binance==1.0.19
python-telegram-bot==20.8